Moved CompoundDataObject to eionet.util package - Refs #25055
