+++ src/main/java/eionet/meta/controllers/ElementFixedValuesController.java
@@ -1,5 +1,6 @@
 package eionet.meta.controllers;
 
+import eionet.util.CompoundDataObject;
 import eionet.meta.application.errors.DuplicateResourceException;
 import eionet.meta.application.errors.UserAuthenticationException;
 import eionet.meta.application.errors.MalformedIdentifierException;
