+++ src/main/java/eionet/meta/controllers/impl/ElementFixedValuesControllerImpl.java
@@ -2,7 +2,7 @@ package eionet.meta.controllers.impl;
 
 import eionet.meta.application.errors.DuplicateResourceException;
 import eionet.meta.controllers.AppContextProvider;
-import eionet.meta.controllers.CompoundDataObject;
+import eionet.util.CompoundDataObject;
 import eionet.meta.controllers.ElementFixedValuesController;
 import eionet.meta.application.errors.MalformedIdentifierException;
 import eionet.meta.application.errors.UserAuthenticationException;
